feat(docs,tests): separate setup/dev docs and reorganize test structure
  - Create DEVELOPMENT.md with comprehensive dev guidelines
  - Move database management, testing, code quality to DEVELOPMENT.md
  - Streamline README.md to focus on quick setup
  - Reorganize tests into tests/unit and tests/integration structure
  - Add test commands: npm run test:unit, test:integration
  - Update Jest config to support new test organization
  - Add unit/integration test commands for professor evaluation
  - Keep tech stack, deployment, resources in README.md
