+++ package.json
@@ -18,7 +18,11 @@
     "db:migrate": "node scripts/run-migrations.js",
     "db:push": "node scripts/push-db.js",
     "test": "jest",
+    "test:unit": "jest tests/unit",
+    "test:integration": "jest tests/integration",
     "test:watch": "jest --watch",
+    "test:watch:unit": "jest tests/unit --watch",
+    "test:watch:integration": "jest tests/integration --watch",
     "test:coverage": "jest --coverage",
     "test:ci": "jest --coverage --watchAll=false --passWithNoTests"
   },
@@ -59,6 +63,7 @@
     "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^30.0.0",
     "@types/node": "^22",
+    "@types/pg": "^8.15.5",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
