+++ tests/unit/components/auth/EmailInput.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import { EmailInput } from '../EmailInput';
+import { EmailInput } from '../../../../src/app/auth/components/EmailInput';
 
 describe('EmailInput Component', () => {
   const mockOnChange = jest.fn();
